Miscwidget refactor
Move classes from MiscWidget into their own modules.
ProgressDialog class is not currently used.
Clean up some pylint issues.

d_rats/map_sources.py:
  Use refactored module.
  Remove log noise for blank lines in source file.

d_rats/miscwidgets.py:
  Move all the classes to their own modules.

d_rats/config.py:
d_rats/formgui.py:
d_rats/geocode_ui.py:
d_rats/map/mapmarkerlist.py:
d_rats/map/mapwindow.py:
d_rats/map/markereditdialog.py:
d_rats/reqobject.py:
  Use refactored modules.

d_rats/keyedlistwidget.py:
d_rats/latlonentry.py:
d_rats/listwidget.py:
d_rats/progressdialog.py:
d_rats/treewidget.py:
  Refactored from miscwidgets.

docs/source/d_rats.rst:
  Updated for new modules.
